Constified all functions, at the cost of memory usage.
Specifically, each call will now perform its own allocation of various
temporary arrays. This is probably safer anyway, as the const-ness
allows for thread-safety and avoids having to deal with
difficult-to-track bugs due to spillover of intermediates between calls.

We could enable use of one-time array allocations with a workspace
class but that's a bit overkill for this function, given that we
don't really expect it to be called lots of times.