Small bugfix for span calculation near start of range.
Previous versions would expand on the left and right edges of the window
in separate if() conditions, resulting in expansion on both sides within
a single loop iteration. This should not happen as expansion should only
occur once per iteration, which ensures that we have the opportunity to
check the span weight. The current bug would make the window slightly
(+1) bigger than it should be... which isn't too bad, but not right.