Switch to regular functions that accept an Options class instance. (#2)
This is simpler and more intuitive than the class methods with parameter
setters; at the cost of not being able to do the chaining of parameter setting.
(Which, it must be said, was probably not a good idea in the first place.)

We removed the in-place compute as it is confusing. Rather, users are expected
to pre-sort their inputs by the x-value. We move the permutation code into its
own class that users can explicitly call to get everything in order.

We streamlined the calculation of the window limits, with a minor bugfix for
the rare case where the distances to the left/right edges are tied.

We streamlined some aspects of the trend fitting code, with a minor bugfix for
the MAD calculation during robustness iterations with non-frequency weights.

We also hardened the tests for various steps, with a whole bunch of new tests
for correct calculation of the windows around each anchor.