Switch to the subpar library for a central parallelization scheme.
This defines the WEIGHTEDLOWESS_CUSTOM_PARALLEL macro by default. It allows us
to get rid of all the macro nonsense and simplifies testing.

We also remove some of the parallelization for the robust calculations, where
the operations are so simple that they are outweighted by the overhead of
parallel setup. We replace them with OpenMP SIMD instructions instead.