Generalize the zero-range check to the robustness iterations.
This allows us to terminate early if we detect that the range of the
non-outlier (i.e., zero-weight) points is zero; the fit is already
perfect so there wouldn't be anything else to do in the next iterations.

It also allows us to define a more appropriate lower bound on the
biweight threshold. Previously, we used the range of the entire dataset,
but this could be skewed by outliers, causing an inflated threshold that
could override a valid MAD estimate. By recomputing the range at each
iteration, we ignore outliers to get a more suitable lower bound.