Upgrade to the latest subpar with no-throw parallelization.
This is possible because none of the parallel sections do any throwing or even
allocations. We also take advantage of the new sanitize_num_workers() function
to improve the setup of the workspaces outside of the parallel sections.

The WEIGHTEDLOWESS_CUSTOM_PARALLEL macro is no longer defined by default.
Rather, we expose the parallelize() function, which (i) provides a central
location for documentation and (ii) avoids problems with macro expansion when
the lambda itself contains more macros, e.g., OpenMP.