Stripped out OpenMP SIMD macros as they're not worth the effort.
Some testing on godbolt indicates that the compiler autovectorizes wherever it
can anyway; the OpenMP macros don't change the compiler's decision much, at
least not in the places that it's used in this codebase. So, we remove them
to reduce the risk of incorrectness and the amount of testing we have to do.

On that note, we can get rid of the OpenMP tests, given that there's no more
macros to test and the parallelization is already handled by subpar's tests.