Shift workspace allocation into the thread to reduce false sharing.
This also requires us to perform an explicit resize() inside
compute_mad() to guarantee that that the vector has enough space, given
that we don't allocate everything up-front anymore. Doing so is probably
safer anyway as it reduces assumptions about the inputs.

Also remove the parallelization arguments for compute_mad(), given
that no parallelization is performed therein.